models/ML/team4-CICD/requirements.txt
# 데이터 처리
pandas==2.1.0
numpy==1.26.0
pyarrow==12.0.1
pyspark==3.5.1

# 머신러닝
scikit-learn==1.3.0

# MLflow
mlflow==2.8.0

# 데이터베이스 연결
SQLAlchemy==2.0.20
pyodbc==4.0.39
pytest==7.4.2
azure-storage-blob>=12.0.0
azureml-core
scipy